pyproject.toml
[project]
name = "colophon"
version = "0.1.0"
description = "Stylometric and linguistic pattern analysis CLI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["stylometry", "linguistics", "nlp", "authorship", "text-analysis"]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
ingest = [
    "pymupdf>=1.24.0",
    "pdfplumber>=0.11.0",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
]
analysis = [
    "spacy>=3.7.0",
    "textstat>=0.7.3",
    "nltk>=3.8.0",
]
report = [
    "plotly>=5.18.0",
    "jinja2>=3.1.0",
]
stylize = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]
ocr = [
    "ocrmypdf>=16.0.0",
]
all = ["colophon[ingest,analysis,report,stylize]"]
dev = [
    "colophon[all]",
    "pytest>=8.0.0",
    "ruff>=0.6.0",
    "pyright>=1.1.0",
]

[project.scripts]
colophon = "colophon.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/colophon"]

[tool.pyright]
include = ["src/colophon"]
venvPath = "."
venv = ".venv"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
